fix(samplefs): Add ros-humble-ros-base for ubuntu22 jammy;Only desktop need firefox-esr

+++ samplefs/make_ubuntu_samplefs.sh
@@ -356,11 +356,13 @@ make_base_root() {
 		chroot "${dst_dir}" /bin/bash -c "ln -sf . /lib/aarch64-none-linux-gnu/11.2.1"
 		chroot "${dst_dir}" /bin/bash -c "ln -sf . /lib/aarch64-none-linux-gnu/11.3.1"
 		
-		# Add firefox-esr
-		chroot "${dst_dir}" /bin/bash -c "apt remove firefox -y"
-		chroot "${dst_dir}" /bin/bash -c "apt install gpg-agent -y"
-		chroot "${dst_dir}" /bin/bash -c "echo -e '\n' | add-apt-repository ppa:mozillateam/ppa"
-		chroot "${dst_dir}" /bin/bash -c "apt install firefox-esr -y"
+		# Ubuntu Desktop Add firefox-esr
+		if [[ $ubuntufs_src == "${LOCAL_DIR}/desktop"  ]] ; then
+			chroot "${dst_dir}" /bin/bash -c "apt remove firefox -y"
+			chroot "${dst_dir}" /bin/bash -c "apt install gpg-agent -y"
+			chroot "${dst_dir}" /bin/bash -c "echo -e '\n' | add-apt-repository ppa:mozillateam/ppa"
+			chroot "${dst_dir}" /bin/bash -c "apt install firefox-esr -y"
+		fi
 	fi
 
 	# upgrade packages
@@ -382,6 +384,11 @@ make_base_root() {
 			fi
 		done
 	fi
+
+	if [ "${RELEASE}" == "jammy" ]; then
+		chroot "${dst_dir}" /bin/bash -c "apt install ros-humble-ros-base -y"
+	fi
+	
 	chroot "${dst_dir}" /bin/bash -c "dpkg --get-selections" | grep -v deinstall | awk '{print $1}' | cut -f1 -d':' | sort > "${tar_file}".info
 
 	chroot "${dst_dir}" /bin/bash -c "pip3 config set global.index-url https://pypi.tuna.tsinghua.edu.cn/simple"
